feat(query): return error response and status code for remote metadata query
