+++ query/src/main/scala/filodb/query/exec/MetadataRemoteExec.scala
@@ -27,13 +27,27 @@ case class MetadataRemoteExec(queryEndpoint: String,
 
   override def sendHttpRequest(execPlan2Span: Span, httpTimeoutMs: Long)
                               (implicit sched: Scheduler): Future[QueryResponse] = {
+    import PromCirceSupport._
+    import io.circe.parser
     remoteExecHttpClient.httpMetadataPost(queryEndpoint, httpTimeoutMs,
       queryContext.submitTime, getUrlParams(), queryContext.traceInfo)
       .map { response =>
-        response.unsafeBody match {
-          case Left(error) =>    // FIXME need to extract statistics from query error, aggregate them, send upstream
-                                 QueryError(queryContext.queryId, QueryStats(), error.error)
-          case Right(successResponse) => toQueryResponse(successResponse, queryContext.queryId, execPlan2Span)
+        // Error response from remote partition is a nested json present in response.body
+        // as response status code is not 2xx
+        if (response.body.isLeft) {
+          parser.decode[ErrorResponse](response.body.left.get) match {
+            case Right(errorResponse) =>
+              QueryError(queryContext.queryId, readQueryStats(errorResponse.queryStats),
+                RemoteQueryFailureException(response.code.toInt, errorResponse.status, errorResponse.errorType,
+                  errorResponse.error))
+            case Left(ex)             => QueryError(queryContext.queryId, QueryStats(), ex)
+          }
+        }
+        else {
+          response.unsafeBody match {
+            case Left(error)            => QueryError(queryContext.queryId, QueryStats(), error.error)
+            case Right(successResponse) => toQueryResponse(successResponse, queryContext.queryId, execPlan2Span)
+          }
         }
       }
   }
